feature(engine): typescript typings
