+++ modules/debug/package.json
@@ -31,8 +31,8 @@
   },
   "dependencies": {
     "@luma.gl/constants": "8.4.0-alpha.1",
-    "@math.gl/core": "^3.4.1",
     "glsl-transpiler": "^1.8.5",
+    "@math.gl/core": "^3.3.0",
     "probe.gl": "^3.2.1",
     "webgl-debug": "^2.0.1"
   },
