+++ modules/engine/package.json
@@ -16,6 +16,7 @@
     "animation",
     "3d"
   ],
+  "types": "src/index.d.ts",
   "main": "dist/es5/index.js",
   "module": "dist/esm/index.js",
   "esnext": "dist/es6/index.js",
@@ -35,7 +36,7 @@
     "@luma.gl/gltools": "8.4.0-alpha.1",
     "@luma.gl/shadertools": "8.4.0-alpha.1",
     "@luma.gl/webgl": "8.4.0-alpha.1",
-    "@math.gl/core": "^3.4.1",
+    "@math.gl/core": "^3.3.0",
     "probe.gl": "^3.2.1"
   },
   "gitHead": "3f8566893e360112486ccf306d3aaa264e072672"
