+++ modules/engine/src/animation/key-frames.d.ts
@@ -1,0 +1,9 @@
+export class KeyFrames {
+  constructor(keyFrames: any);
+  setKeyFrames(keyFrames: any): void;
+  setTime(time: any): void;
+  getStartTime(): any;
+  getEndTime(): any;
+  getStartData(): any;
+  getEndData(): any;
+}
