+++ modules/engine/src/animation/timeline.d.ts
@@ -1,0 +1,14 @@
+export class Timeline {
+  constructor();
+  addChannel(props: any): number;
+  removeChannel(handle: any): void;
+  isFinished(handle: any): boolean;
+  getTime(handle: any): any;
+  setTime(time: any): void;
+  play(): void;
+  pause(): void;
+  reset(): void;
+  attachAnimation(animation: any, channelHandle: any): number;
+  detachAnimation(handle: any): void;
+  update(engineTime: any): void;
+}
