+++ modules/engine/src/geometries/cone-geometry.d.ts
@@ -1,0 +1,5 @@
+import TruncatedConeGeometry from "./truncated-cone-geometry";
+
+export default class ConeGeometry extends TruncatedConeGeometry {
+  constructor(props?: {});
+}
