+++ modules/engine/src/geometries/cube-geometry.d.ts
@@ -1,0 +1,5 @@
+import Geometry from "../geometry/geometry";
+
+export default class CubeGeometry extends Geometry {
+  constructor(props?: {});
+}
