+++ modules/engine/src/geometries/cylinder-geometry.d.ts
@@ -1,0 +1,5 @@
+import TruncatedConeGeometry from "./truncated-cone-geometry";
+
+export default class CylinderGeometry extends TruncatedConeGeometry {
+  constructor(props?: {});
+}
