+++ modules/engine/src/geometries/ico-sphere-geometry.d.ts
@@ -1,0 +1,5 @@
+import Geometry from "../geometry/geometry";
+
+export default class IcoSphereGeometry extends Geometry {
+  constructor(props?: {});
+}
