+++ modules/engine/src/geometries/plane-geometry.d.ts
@@ -1,0 +1,5 @@
+import Geometry from "../geometry/geometry";
+
+export default class PlaneGeometry extends Geometry {
+  constructor(props?: {});
+}
