+++ modules/engine/src/geometries/sphere-geometry.d.ts
@@ -1,0 +1,5 @@
+import Geometry from "../geometry/geometry";
+
+export default class SphereGeometry extends Geometry {
+  constructor(props?: {});
+}
