+++ modules/engine/src/geometries/truncated-cone-geometry.d.ts
@@ -1,0 +1,5 @@
+import Geometry from "../geometry/geometry";
+
+export default class TruncatedConeGeometry extends Geometry {
+  constructor(props?: {});
+}
