+++ modules/engine/src/geometry/geometry-utils.d.ts
@@ -1,0 +1,1 @@
+export function unpackIndexedGeometry(geometry: any): any;
