+++ modules/engine/src/lib/model-utils.d.ts
@@ -1,0 +1,10 @@
+export function getBuffersFromGeometry(
+	gl: WebGLRenderingContext,
+	geometry: any,
+	options?: any
+): {};
+
+export function inferAttributeAccessor(
+	attributeName: any,
+	attribute: any
+): void;
