+++ modules/engine/src/utils/clip-space.d.ts
@@ -1,0 +1,5 @@
+import Model from "../lib/model";
+
+export default class ClipSpace extends Model {
+  constructor(gl: WebGLRenderingContext, opts: any);
+}
