+++ modules/test-utils/package.json
@@ -16,6 +16,7 @@
     "automation",
     "testing"
   ],
+  "types": "src/index.d.ts",
   "main": "dist/es5/index.js",
   "module": "dist/esm/index.js",
   "esnext": "dist/es6/index.js",
@@ -31,10 +32,10 @@
     "probe.gl": "^3.2.1"
   },
   "peerDependencies": {
-    "@luma.gl/core": "8.4.0-alpha.1",
-    "@luma.gl/debug": "8.4.0-alpha.1",
-    "@luma.gl/gltools": "8.4.0-alpha.1",
-    "@luma.gl/webgl": "8.4.0-alpha.1",
+    "@luma.gl/core": "^8.1.0",
+    "@luma.gl/debug": "^8.1.0",
+    "@luma.gl/gltools": "^8.1.0",
+    "@luma.gl/webgl": "^8.1.0",
     "@probe.gl/test-utils": "^3.2.1"
   },
   "gitHead": "3f8566893e360112486ccf306d3aaa264e072672"
