+++ modules/test-utils/src/context.d.ts
@@ -1,0 +1,4 @@
+export function createTestContext(options?: object): WebGLRenderingContext;
+
+  // Create headless gl context (for running under Node.js)
+export function createHeadlessContext(options?: object): object;
