+++ modules/test-utils/src/index.d.ts
@@ -1,0 +1,3 @@
+export {default as SnapshotTestRunner} from './snapshot-test-runner';
+export {default as PerformanceTestRunner} from './performance-test-runner';
+export {createTestContext, createHeadlessContext} from './context';
