+++ modules/test-utils/src/performance-test-runner.d.ts
@@ -1,0 +1,5 @@
+import TestRunner, {TestRunnerOptions} from './test-runner';
+
+export default class PerformanceTestRunner extends TestRunner {
+  constructor(props: TestRunnerOptions);
+}
