+++ modules/test-utils/src/performance-test-runner.js
@@ -21,6 +21,7 @@ export default class PerformanceTestRunner extends TestRunner {
     this._fps.timeEnd();
     this._fps.timeStart();
 
+    // @ts-ignore
     if (this._fps.count > this.testOptions.maxFramesToRender) {
       animationProps.done();
     }
@@ -29,6 +30,7 @@ export default class PerformanceTestRunner extends TestRunner {
   }
 
   assert(testCase) {
+    // @ts-ignore
     const targetFPS = testCase.targetFPS || this.testOptions.targetFPS;
     const count = this._fps.count;
     const fps = this._fps.getHz();
