+++ modules/test-utils/src/snapshot-test-runner.d.ts
@@ -1,0 +1,5 @@
+import TestRunner, {TestRunnerOptions} from './test-runner';
+
+export default class SnapshotTestRunner extends TestRunner {
+  constructor(props: TestRunnerOptions);
+}
